requirements.txt
appdirs==1.4.2
cycler==0.10.0
matplotlib==2.0.0
numpy==1.12.0
olefile==0.44
packaging==16.8
pandas==0.19.2
Pillow==4.0.0
pygame==1.9.3
pyparsing==2.1.10
pytagcloud==0.3.5
python-dateutil==2.6.0
pytz==2016.10
simplejson==3.10.0
six==1.10.0
